+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## 0.2.0 - 2026-07-29
+
 ### Added
 
 - Added optional `nonce` attribute support to `JsonLdRenderer` and `render_schema_org` Twig function for Content Security Policy (CSP) compliance.